artificial_intelligence/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
nltk>=3.8.1
textstat>=0.7.3
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
moviepy>=1.0.3
python-docx>=0.8.11
